Do not create the list of files multiple times
This is a performance optimization to avoid expressions like "files.empty?
|| files.nil?" like in "project.rb" to trigger the creation of the list of
files multiple times.

As a side effect, this fixes a bug where calling FSProject.files with a
single file for the second time would return all files in the directory
instaed of that single file. This is because in case of a single file the
implementation overwrites @path with the directory of the single file.
The next time FSProject.files is called, @path is not a file anymore, and
the pattern will be set to "*", matching all files.