fs_project: Fix for calling "files" with a file multiple times
This fixes a bug where calling FSProject.files with a single file for the
second time would return all files in the directory instead of that single
file. This is because in case of a single file the original implementation
would overwrite @path with the directory of the single file. The next time
FSProject.files is called, @path is not a file anymore, and the pattern
will be set to "*", matching all files.

Fix this by splitting the path into directory and pattern, and always
constructing the path from these.