Complete support for R/CRAN License conventions
This continues work begun with #129. In particular it allows fallback to
the readme and project matchers if the license_file does not contain a
valid license.

We also special case the most common naming conventions with CRAN
packages, as the license matching expects exact name matching, so
without this most of the ly used license terms in R packages would not be
found. The top 6 most common strings on CRAN (with counts) are

3484 GPL (>= 2)
1973 GPL-2
1159 GPL-3
534 MIT + file LICENSE
522 GPL
459 GPL (>= 3)

With these changes many more R package licenses should be properly
detected by Licensee.