matchers/cargo: Allow SPDX expressions
Using a regex that is similar to the one in
lib/licensee/matchers/spdx.rb in allowing spaces and parens.  Parens
aren't currently supported by Cargo's crates.io backing [1,2], but
that's hopefully a temporary limitation.  And Cargo/crates.io already
supports OR forms [3].

Ideally licensee would return the full license expression in these
cases instead of 'other' [4].  But since that conversion is happening
outside the matcher, this commit isn't making that any worse (although
it means this test isn't quite a unit test).

[1]: https://github.com/rust-lang-nursery/license-exprs/issues/3
[2]: https://github.com/rust-lang/cargo/pull/4898/commits/c89dd6452ce71765eb47c675d32ecc3834f78691
[3]: https://github.com/rust-lang/cargo/pull/4920
[4]: https://github.com/benbalter/licensee/issues/244