gemspec: Allow use of Rugged v1.0
Rugged v1.0 was released some time ago, wrapping the first stable major
release of libgit2. While no breaking changes should exist, the major
version bump was performed to indicate that libgit2 is now considered
API-stable across all of v1.X.

As the Gemspec is currently specifying Rugged's version to be '~> 0.24',
we're not pulling in the new version. Let's bump this to allow
everything between `[v0.24, v2.0[`.