Punish length differences more, give fill-in fields more leeway
Also follow up on 5436184d822c1cb9dbad8cb574cb0c306c35bddc by
adding - (in addition to /) as a word character so to minimize
wordset differences when one adds a whole URL with -'s in it to
a fill in (would also work for hyphenated names)

Overall this seems right, though it'd be comforting to have a far
larger test corpus. Length shouldn't vary -- licenses shouldn't be
modified -- but it's inevitable with licenses with fill-in fields.

Fixes #474 which was about the former -- a modified Apache 2.0 license.