Fine tune license filenames
Exclude .go or .xml extensions for files with more than just 'license' as suffixless basename, so that license-checks.xml is excluded (fixes #479) but license.xml is still included as if it occurred we can't confidently say it isn't a license.

Limit exclusion of .spdx and .header to files with only 'license' as suffixless basename, because only license.spdx and license.header are known in the wild files that should be excluded.

Allow more