Tighten license equality comparison
Previously if you attempted to compare `Licensee::License` to a
boolean, it would fail with this obscure error:

```
NoMethodError Exception: undefined method `key' for true:TrueClass
```

This happened when Rails `ActiveSupport::Cache::Entry#dup_value!`
attempted to duplicate the `License::License` by comparing it to
boolean values.

We fix this by checking that `Licensee::License` is the other model.