Add support for REUSE-style `LICENSES/` directories
[reuse-tool] is becoming a popular way to manage licenses, and is now
used by projects like [Rust] and [curl]. License files must be placed
in a [LICENSES/] directory. This patch simply teaches the `licensee`
tool to also search in this directory for licenses.

Along the way, since this changes the output to include the directory
by implementing an old TODO to get `Project_file.path` to be the
absolute path to the project root, migrates `detect.rb` over to using it.

Closes #737

[reuse-tool]: https://reuse.software/
[Rust]: https://github.com/rust-lang/rust/tree/master/LICENSES
[curl]: https://github.com/curl/curl/tree/master/LICENSES
[LICENSES/]: https://reuse.software/spec-3.3/#license-files